refactor: refactor according mentors check
